+++ test/main/Main.hs
@@ -15,6 +15,7 @@ import Test.Calendar.Valid
 import Test.Calendar.Week
 import Test.Calendar.Year
 import Test.Clock.Conversion
+import Test.Clock.Pattern
 import Test.Clock.Resolution
 import Test.Clock.TAI
 import Test.Format.Compile ()
@@ -48,7 +49,7 @@ tests =
             , testYear
             , testDuration
             ]
-        , testGroup "Clock" [testClockConversion, testResolutions, testTAI]
+        , testGroup "Clock" [testClockPatterns, testClockConversion, testResolutions, testTAI]
         , testGroup "Format" [testFormat, testParseTime, testISO8601]
         , testGroup "LocalTime" [testTime, testTimeOfDay, testCalendarDiffTime]
         ]